backend/backend/backend/requirements.txt
django
django-rest-framework
django-cors-headers
torch
torchvision
opencv-python
facenet-pytorch
Pillow